requirements.txt
google-cloud-bigquery[pandas]>=3.25.0
pandas>=2.2.0
requests>=2.32.0
